Remove and update
I removed the inline CSS from images and paragraphs. Having the width and height of items in the HTML will hinder you from easily updating items and styles. All of that should be in the CSS by placing classes on elements and defining styles in the CSS. 
You can target all paragraphs, place a container around elements to give them all the background color as I did with the header container, and target images within the images-on-right or left to remove the margin-top so the text and images are aligned to the top.
I also added a closing div where needed and it fixed the footer.
Great work, Lee!